server/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.7
python-dotenv==1.0.0
Flask-CORS==4.0.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
gspread==5.12.0
openai>=1.35.0,<2.0.0
sentence-transformers>=2.7.0,<3.0.0
huggingface_hub>=0.26.0,<1.0.0
faiss-cpu==1.7.4
nltk==3.8.1
numpy<2.0.0,>=1.21.0
pytz==2023.3
